컷씬 기본기능 작업중.

+++ fe_client/Assets/Script/01.Battle/01.Entity/01.Status/UnitStatus.cs
@@ -19,12 +19,12 @@ namespace Battle
 
         public int GetPoint(EnumUnitPoint _point_type)
         {
-            return m_point_repository.GetValue((int)_point_type);
+            return (int)m_point_repository.GetValue((int)_point_type);
         }
 
         public int GetStatus(EnumUnitStatus _status_type)
         {
-            return m_status_repository.GetValue((int)_status_type);
+            return (int)m_status_repository.GetValue((int)_status_type);
         }
 
         public bool HasAttribute(EnumUnitAttribute _attribute_type)
